+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG for FlatCAM beta
 - in Gerber Editor - modified the UI
 - in Gerber Editor - added new parameters for selected polygons: is_valid and area
 - in Gerber Editor - for Disc and SemiDisc tools now it is allowed to start without a defined aperture; it will be assumed that we work with an aperture with size zero
+- Subtraction Tool - fixed issue when from a given polygon it is subtracted multiple ones, the result being that the subtraction is not done
 
 2.12.2020